Modified delete all logic in native data store

+++ MicroFrameworkPK_v4_3/Samraksh/Interops/Samraksh_SPOT/Samraksh_SPOT/NativeStubs/Samraksh_SPOT_Samraksh_SPOT_NonVolatileMemory_DataStore.h
@@ -38,8 +38,8 @@ namespace Samraksh
                 static UINT16 ReadAllDataIDs( CLR_RT_HeapBlock* pMngObj, CLR_RT_TypedArray_UINT32 param0, HRESULT &hr );
                 static INT8 Delete( CLR_RT_HeapBlock* pMngObj, UINT32 param0, HRESULT &hr );
                 static INT32 Write( CLR_RT_HeapBlock* pMngObj, UINT32 param0, CLR_RT_TypedArray_UINT8 param1, UINT32 param2, INT32 param3, HRESULT &hr );
-                static INT8 DeleteAll( HRESULT &hr );
-                static INT8 DataStoreGC( HRESULT &hr );
+                static INT32 DeleteAll( HRESULT &hr );
+                static INT32 DataStoreGC( HRESULT &hr );
                 static INT8 GetReadWriteStatus( HRESULT &hr );
                 static INT32 GetLastErrorStatus( HRESULT &hr );
             };
